Changes how clients handle port selection
Clients now don't have a static port that they have to connect to, it is dynamically allocated within a range of ports, based on which port is open at the time the client wants to connect. This lifts the restriction of client 1 having to join before client 2, however I am unsure of how stable and how well this amendment to the code works  for both connecting to the server and other issues that may pop up in the code in general. For now i'll branch this change until I can further investigate the effect theses changes will have.