Add files via upload
Fixed issue causing the server to terminate connection to client as soon as they loaded up. Before hand i would connect to the server to see if that port was open, turns out that would then think that was the client loading up I would then close it return the port and use it to connect [very big brain]. Now once a port has been connected to I return the socket connection and use that. Found loads of issues in the main code which are not fun client 2 gets a "ConnectionResetError: [Errno 54] Connection reset by peer" when it is waiting for response from the sever to refresh board. The server doesn't wait for clients input which causes a "_pickle.UnpicklingError: unpickling stack underflow" error. Need to investigate and refactor the base server code. HOW FUN !